+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added basic auto walk
 - Added weapon/attack sound effects
 - M2 attack for bow
+- Hotbar persistence.
 
 ### Changed